LPS-54098 Move test package to the package.test package. Two OSGi modules can't share the same package.
