+++ modules/apps/bookmarks/bookmarks-test/test/integration/com/liferay/bookmarks/service/test/BookmarksEntryLocalServiceTest.java
@@ -12,32 +12,40 @@
  * details.
  */
 
-package com.liferay.bookmarks.service;
+package com.liferay.bookmarks.service.test;
 
 import com.liferay.bookmarks.model.BookmarksEntry;
+import com.liferay.bookmarks.service.BookmarksEntryLocalServiceUtil;
 import com.liferay.bookmarks.util.test.BookmarksTestUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
+import com.liferay.portal.kernel.test.rule.Sync;
+import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
-
+import com.liferay.portlet.asset.model.AssetEntry;
+import com.liferay.portlet.asset.service.AssetEntryLocalServiceUtil;
+import java.util.List;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 
 /**
- * @author Brian Wing Shun Chan
+ * @author Michael C. Han
  */
-public class BookmarksEntryServiceTest {
+@Sync
+public class BookmarksEntryLocalServiceTest {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
 		new AggregateTestRule(
-			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE);
+			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE,
+			SynchronousDestinationTestRule.INSTANCE);
 
 	@Before
 	public void setUp() throws Exception {
@@ -45,24 +53,22 @@ public class BookmarksEntryServiceTest {
 	}
 
 	@Test
-	public void testAddEntry() throws Exception {
-		BookmarksTestUtil.addEntry(_group.getGroupId(), true);
-	}
-
-	@Test
-	public void testDeleteEntry() throws Exception {
+	public void testGetNoAssetFolders() throws Exception {
 		BookmarksEntry entry = BookmarksTestUtil.addEntry(
 			_group.getGroupId(), true);
 
-		BookmarksEntryServiceUtil.deleteEntry(entry.getEntryId());
-	}
+		AssetEntry assetEntry = AssetEntryLocalServiceUtil.fetchEntry(
+			BookmarksEntry.class.getName(), entry.getEntryId());
 
-	@Test
-	public void testGetEntry() throws Exception {
-		BookmarksEntry entry = BookmarksTestUtil.addEntry(
-			_group.getGroupId(), true);
+		Assert.assertNotNull(assetEntry);
 
-		BookmarksEntryServiceUtil.getEntry(entry.getEntryId());
+		AssetEntryLocalServiceUtil.deleteAssetEntry(assetEntry);
+
+		List<BookmarksEntry> bookmarksEntries =
+			BookmarksEntryLocalServiceUtil.getNoAssetEntries();
+
+		Assert.assertEquals(1, bookmarksEntries.size());
+		Assert.assertEquals(entry, bookmarksEntries.get(0));
 	}
 
 	@DeleteAfterTestRun
