+++ modules/apps/bookmarks/bookmarks-test/test/integration/com/liferay/bookmarks/service/test/BookmarksEntryServiceTest.java
@@ -12,41 +12,33 @@
  * details.
  */
 
-package com.liferay.bookmarks.service;
+package com.liferay.bookmarks.service.test;
 
 import com.liferay.bookmarks.model.BookmarksEntry;
+import com.liferay.bookmarks.service.BookmarksEntryServiceUtil;
 import com.liferay.bookmarks.util.test.BookmarksTestUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
-import com.liferay.portal.kernel.test.rule.Sync;
-import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
-import com.liferay.portlet.asset.model.AssetEntry;
-import com.liferay.portlet.asset.service.AssetEntryLocalServiceUtil;
 
-import java.util.List;
-
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 
 /**
- * @author Michael C. Han
+ * @author Brian Wing Shun Chan
  */
-@Sync
-public class BookmarksEntryLocalServiceTest {
+public class BookmarksEntryServiceTest {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
 		new AggregateTestRule(
-			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE,
-			SynchronousDestinationTestRule.INSTANCE);
+			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE);
 
 	@Before
 	public void setUp() throws Exception {
@@ -54,22 +46,24 @@ public class BookmarksEntryLocalServiceTest {
 	}
 
 	@Test
-	public void testGetNoAssetFolders() throws Exception {
+	public void testAddEntry() throws Exception {
+		BookmarksTestUtil.addEntry(_group.getGroupId(), true);
+	}
+
+	@Test
+	public void testDeleteEntry() throws Exception {
 		BookmarksEntry entry = BookmarksTestUtil.addEntry(
 			_group.getGroupId(), true);
 
-		AssetEntry assetEntry = AssetEntryLocalServiceUtil.fetchEntry(
-			BookmarksEntry.class.getName(), entry.getEntryId());
+		BookmarksEntryServiceUtil.deleteEntry(entry.getEntryId());
+	}
 
-		Assert.assertNotNull(assetEntry);
+	@Test
+	public void testGetEntry() throws Exception {
+		BookmarksEntry entry = BookmarksTestUtil.addEntry(
+			_group.getGroupId(), true);
 
-		AssetEntryLocalServiceUtil.deleteAssetEntry(assetEntry);
-
-		List<BookmarksEntry> bookmarksEntries =
-			BookmarksEntryLocalServiceUtil.getNoAssetEntries();
-
-		Assert.assertEquals(1, bookmarksEntries.size());
-		Assert.assertEquals(entry, bookmarksEntries.get(0));
+		BookmarksEntryServiceUtil.getEntry(entry.getEntryId());
 	}
 
 	@DeleteAfterTestRun
