+++ modules/apps/bookmarks/bookmarks-test/test/integration/com/liferay/bookmarks/subscription/test/BookmarksSubscriptionBaseModelTest.java
@@ -12,24 +12,27 @@
  * details.
  */
 
-package com.liferay.bookmarks.subscription;
+package com.liferay.bookmarks.subscription.test;
 
-import com.liferay.bookmarks.constants.BookmarksConstants;
-import com.liferay.bookmarks.constants.BookmarksPortletKeys;
 import com.liferay.bookmarks.model.BookmarksEntry;
+import com.liferay.bookmarks.model.BookmarksFolder;
 import com.liferay.bookmarks.service.BookmarksEntryLocalServiceUtil;
 import com.liferay.bookmarks.service.BookmarksFolderLocalServiceUtil;
 import com.liferay.bookmarks.util.test.BookmarksTestUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
+import com.liferay.portal.kernel.test.util.RoleTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.util.Constants;
+import com.liferay.portal.model.ResourceConstants;
+import com.liferay.portal.model.RoleConstants;
+import com.liferay.portal.security.permission.ActionKeys;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
 import com.liferay.portal.test.rule.SynchronousMailTestRule;
-import com.liferay.portlet.subscriptions.test.BaseSubscriptionLocalizedContentTestCase;
+import com.liferay.portlet.subscriptions.test.BaseSubscriptionBaseModelTestCase;
 
 import org.junit.ClassRule;
 import org.junit.Rule;
@@ -38,8 +41,8 @@ import org.junit.Rule;
  * @author Roberto Díaz
  */
 @Sync
-public class BookmarksSubscriptionLocalizedContentTest
-	extends BaseSubscriptionLocalizedContentTestCase {
+public class BookmarksSubscriptionBaseModelTest
+	extends BaseSubscriptionBaseModelTestCase {
 
 	@ClassRule
 	@Rule
@@ -68,31 +71,39 @@ public class BookmarksSubscriptionLocalizedContentTest
 	}
 
 	@Override
-	protected void addSubscriptionContainerModel(long containerModelId)
+	protected long addContainerModel(long userId, long containerModelId)
 		throws Exception {
 
-		BookmarksFolderLocalServiceUtil.subscribeFolder(
-			user.getUserId(), group.getGroupId(), containerModelId);
+		ServiceContext serviceContext =
+			ServiceContextTestUtil.getServiceContext(
+				group.getGroupId(), userId);
+
+		_folder = BookmarksFolderLocalServiceUtil.addFolder(
+			userId, containerModelId, RandomTestUtil.randomString(),
+			RandomTestUtil.randomString(), serviceContext);
+
+		return _folder.getFolderId();
 	}
 
 	@Override
-	protected String getPortletId() {
-		return BookmarksPortletKeys.BOOKMARKS;
+	protected void addSubscriptionBaseModel(long baseModelId) throws Exception {
+		BookmarksEntryLocalServiceUtil.subscribeEntry(
+			user.getUserId(), baseModelId);
 	}
 
 	@Override
-	protected String getServiceName() {
-		return BookmarksConstants.SERVICE_NAME;
-	}
+	protected void removeContainerModelResourceViewPermission()
+		throws Exception {
 
-	@Override
-	protected String getSubscriptionAddedBodyPreferenceName() {
-		return "emailEntryAddedBody";
-	}
+		RoleTestUtil.removeResourcePermission(
+			RoleConstants.GUEST, BookmarksFolder.class.getName(),
+			ResourceConstants.SCOPE_INDIVIDUAL,
+			String.valueOf(_folder.getFolderId()), ActionKeys.VIEW);
 
-	@Override
-	protected String getSubscriptionUpdatedBodyPreferenceName() {
-		return "emailEntryUpdatedBody";
+		RoleTestUtil.removeResourcePermission(
+			RoleConstants.SITE_MEMBER, BookmarksFolder.class.getName(),
+			ResourceConstants.SCOPE_INDIVIDUAL,
+			String.valueOf(_folder.getFolderId()), ActionKeys.VIEW);
 	}
 
 	@Override
@@ -114,5 +125,7 @@ public class BookmarksSubscriptionLocalizedContentTest
 			RandomTestUtil.randomString(), entry.getUrl(),
 			entry.getDescription(), serviceContext);
 	}
+
+	private BookmarksFolder _folder;
 
 }
