+++ modules/apps/bookmarks/bookmarks-test/test/integration/com/liferay/bookmarks/verify/test/BookmarksServiceVerifyProcessTest.java
@@ -12,13 +12,14 @@
  * details.
  */
 
-package com.liferay.bookmarks.verify;
+package com.liferay.bookmarks.verify.test;
 
 import com.liferay.bookmarks.model.BookmarksEntry;
 import com.liferay.bookmarks.model.BookmarksFolder;
 import com.liferay.bookmarks.service.BookmarksEntryLocalServiceUtil;
 import com.liferay.bookmarks.service.BookmarksFolderLocalServiceUtil;
 import com.liferay.bookmarks.util.test.BookmarksTestUtil;
+import com.liferay.bookmarks.verify.BookmarksServiceVerifyProcess;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
